Added repository and custom data validation.

+++ Repository/PlayerRepository.cs
@@ -7,7 +7,7 @@ namespace Gc_Broadcasting_Api.Repository;
 public sealed class PlayerRepository(Database database) : IPlayerRepo {
     private readonly Database _database = database;
 
-    public async Task<bool> CreateNewPlayer(Player player, CancellationToken ct) {
+    public async Task<bool> CreatePlayer(Player player, CancellationToken ct = default) {
         ct.ThrowIfCancellationRequested();
 
         if (player is null) { return false; }
@@ -21,13 +21,15 @@ public sealed class PlayerRepository(Database database) : IPlayerRepo {
         } 
     }
 
-    public async Task<bool> DeletePlayer(string playerId, CancellationToken ct) {
+    public async Task<bool> DeletePlayer(string playerId, CancellationToken ct = default) {
         ct.ThrowIfCancellationRequested();
 
         if (playerId is null) { return false; }
 
+        var filter = Builders<Player>.Filter.Eq(r => r.Id, playerId);
+        if (filter is null) { return false; }
+
         try {
-            var filter = Builders<Player>.Filter.Eq(r => r.Id, playerId);
             await _database.Players.DeleteOneAsync(filter, null, ct);
             return true;
         }
@@ -36,13 +38,15 @@ public sealed class PlayerRepository(Database database) : IPlayerRepo {
         }
     }
 
-    public async Task<Player> GetPlayerByName(string name, CancellationToken ct) {
+    public async Task<Player> GetPlayer(string name, CancellationToken ct = default) {
         ct.ThrowIfCancellationRequested();
 
         ArgumentNullException.ThrowIfNull(name);
 
+        var filter = Builders<Player>.Filter.Eq(x => x.Name!, name);
+        if (filter is null) { return new Player { }; }
+
         try {
-            var filter = Builders<Player>.Filter.Eq(x => x.Name!, name);
             return await _database.Players.Find(filter).FirstOrDefaultAsync(ct);
         }
         catch (Exception){
@@ -50,31 +54,33 @@ public sealed class PlayerRepository(Database database) : IPlayerRepo {
         }
     }
 
-    public async Task<IEnumerable<Player>> GetPlayersByTeam(int teamId, CancellationToken ct) {
+    public async Task<IEnumerable<Player>> GetPlayers(int teamId, CancellationToken ct = default) {
         ct.ThrowIfCancellationRequested();
 
         if(teamId <= 0) { return Enumerable.Empty<Player>(); }
 
+        var filter = Builders<Player>.Filter.Eq(x => x.TeamId, teamId);
+        if (filter is null) { return Enumerable.Empty<Player>(); } 
+
         try {
-            var filter = Builders<Player>.Filter.Eq(x => x.TeamId, teamId);
             var res = await _database.Players.FindAsync(filter, null, ct);
             return await res.ToListAsync(ct);
         }
         catch (Exception){ return Enumerable.Empty<Player>(); }
     }
 
-    public async Task<bool> UpdatePlayerDetails(Player newPlayerDetails, CancellationToken ct) {
+    public async Task<bool> UpdatePlayer(Player newPlayerDetails, CancellationToken ct = default) {
         ct.ThrowIfCancellationRequested();
 
         ArgumentNullException.ThrowIfNull(newPlayerDetails);
 
+        var filter = Builders<Player>.Filter.Eq(p => p.Id, newPlayerDetails.Id);
+        if (filter is null) { return false; }
+
+        var oldPlayerDetails = await _database.Players.Find(filter).FirstOrDefaultAsync(ct);
+        if (oldPlayerDetails is null) { return false; }
+
         try {
-            var filter = Builders<Player>.Filter.Eq(p => p.Id, newPlayerDetails.Id);
-            var oldPlayerDetails = await _database.Players.Find(filter).FirstOrDefaultAsync(ct);
-            if (oldPlayerDetails is null) {
-                ArgumentNullException.ThrowIfNull(oldPlayerDetails);
-                return false; 
-            }
             var updateCondition = Builders<Player>.Update
                 .Set(u => u.Name, newPlayerDetails.Name)
                 .Set(u => u.Id, newPlayerDetails.Id)
